break out filters as separate component
