+++ src/pages/workspaces/List.ts
@@ -1,0 +1,10 @@
+import { WorkspacesList } from 'src/pages/workspaces/workspacesList/WorkspacesList';
+
+export const navPaths = [
+  {
+    name: 'workspaces',
+    path: '/workspaces',
+    component: WorkspacesList,
+    title: 'Workspaces',
+  },
+];
