+++ src/pages/workspaces/WorkspacesList/WorkspacesList.ts
@@ -5,10 +5,9 @@ import { div, h, p, span } from 'react-hyperscript-helpers';
 import { AutoSizer } from 'react-virtualized';
 import { CloudProviderIcon } from 'src/components/CloudProviderIcon';
 import Collapse from 'src/components/Collapse';
-import { Link, Select, topSpinnerOverlay, transparentSpinnerOverlay } from 'src/components/common';
+import { Link, topSpinnerOverlay, transparentSpinnerOverlay } from 'src/components/common';
 import FooterWrapper from 'src/components/FooterWrapper';
 import { icon } from 'src/components/icons';
-import { DelayedSearchInput } from 'src/components/input';
 import LeaveResourceModal from 'src/components/LeaveResourceModal';
 import { FirstParagraphMarkdownViewer } from 'src/components/markdown';
 import NewWorkspaceModal from 'src/components/NewWorkspaceModal';
@@ -23,7 +22,6 @@ import {
   useWorkspaces,
   WorkspaceStarControl,
   WorkspaceSubmissionStatusIcon,
-  WorkspaceTagSelect,
 } from 'src/components/workspace-utils';
 import { Ajax } from 'src/libs/ajax';
 import { isAzureUser } from 'src/libs/auth';
@@ -37,7 +35,6 @@ import { authStore } from 'src/libs/state';
 import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
 import {
-  cloudProviderLabels,
   cloudProviderTypes,
   getCloudProviderFromWorkspace,
   WorkspaceAccessLevels,
@@ -51,6 +48,8 @@ import LockWorkspaceModal from 'src/pages/workspaces/workspace/LockWorkspaceModa
 import { RequestAccessModal } from 'src/pages/workspaces/workspace/RequestAccessModal';
 import ShareWorkspaceModal from 'src/pages/workspaces/workspace/ShareWorkspaceModal/ShareWorkspaceModal';
 import WorkspaceMenu from 'src/pages/workspaces/workspace/WorkspaceMenu';
+
+import { WorkspaceFilters } from './WorkspaceFilters';
 
 const styles = {
   tableCellContainer: {
@@ -144,7 +143,7 @@ interface WorkspaceSort {
   direction: 'desc' | 'asc';
 }
 
-export const WorkspaceList = () => {
+export const WorkspacesList: React.FC<{}> = () => {
   const {
     workspaces,
     refresh: refreshWorkspaces,
@@ -563,79 +562,7 @@ export const WorkspaceList = () => {
             ]),
           ]
         ),
-      div({ style: { display: 'flex', margin: '1rem 0' } }, [
-        div({ style: { ...styles.filter, flexGrow: 1.5 } }, [
-          h(DelayedSearchInput, {
-            placeholder: 'Search by keyword',
-            'aria-label': 'Search workspaces by keyword',
-            onChange: (newFilter) => Nav.updateSearch({ ...query, filter: newFilter || undefined }),
-            value: filter,
-          }),
-        ]),
-        div({ style: styles.filter }, [
-          h(WorkspaceTagSelect, {
-            isClearable: true,
-            isMulti: true,
-            formatCreateLabel: _.identity,
-            value: _.map((tag) => ({ label: tag, value: tag }), tagsFilter),
-            placeholder: 'Tags',
-            'aria-label': 'Filter by tags',
-            onChange: (data) => Nav.updateSearch({ ...query, tagsFilter: _.map('value', data) }),
-          }),
-        ]),
-        div({ style: styles.filter }, [
-          h(Select<string, true>, {
-            isClearable: true,
-            isMulti: true,
-            isSearchable: false,
-            placeholder: 'Access levels',
-            'aria-label': 'Filter by access levels',
-            value: accessLevelsFilter,
-            onChange: (data) => Nav.updateSearch({ ...query, accessLevelsFilter: _.map('value', data) }),
-            options: [...workspaceAccessLevels], // need to re-create the list otherwise the readonly type of workspaceAccessLevels conflicts with the type of options
-            getOptionLabel: ({ value }) => Utils.normalizeLabel(value),
-          }),
-        ]),
-        div({ style: styles.filter }, [
-          h(Select<string, false>, {
-            isClearable: true,
-            isMulti: false,
-            placeholder: 'Billing project',
-            'aria-label': 'Filter by billing project',
-            value: projectsFilter,
-            hideSelectedOptions: true,
-            onChange: (data) => Nav.updateSearch({ ...query, projectsFilter: data?.value || undefined }),
-            options: _.flow(_.map('workspace.namespace'), _.uniq, _.sortBy(_.identity))(workspaces),
-          }),
-        ]),
-        div({ style: styles.filter }, [
-          h(Select<string, true>, {
-            isClearable: true,
-            isMulti: true,
-            isSearchable: false,
-            placeholder: 'Submission status',
-            'aria-label': 'Filter by submission status',
-            value: submissionsFilter,
-            hideSelectedOptions: true,
-            onChange: (data) => Nav.updateSearch({ ...query, submissionsFilter: _.map('value', data) }),
-            options: ['running', 'success', 'failure'],
-            getOptionLabel: ({ value }) => Utils.normalizeLabel(value),
-          }),
-        ]),
-        div({ style: { ...styles.filter, marginRight: 0 } }, [
-          h(Select<string>, {
-            isClearable: true,
-            isMulti: false,
-            placeholder: 'Cloud platform',
-            'aria-label': 'Filter by cloud platform',
-            value: cloudPlatformFilter,
-            hideSelectedOptions: true,
-            onChange: (data) => Nav.updateSearch({ ...query, cloudPlatform: data?.value || undefined }),
-            options: _.sortBy((cloudProvider) => cloudProviderLabels[cloudProvider], _.keys(cloudProviderTypes)),
-            getOptionLabel: ({ value }) => cloudProviderLabels[value],
-          }),
-        ]),
-      ]),
+      h(WorkspaceFilters, { workspaces }),
       h(
         SimpleTabBar,
         {
@@ -697,12 +624,3 @@ export const WorkspaceList = () => {
     ]),
   ]);
 };
-
-export const navPaths = [
-  {
-    name: 'workspaces',
-    path: '/workspaces',
-    component: WorkspaceList,
-    title: 'Workspaces',
-  },
-];
